requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Web Scraping
playwright==1.40.0
beautifulsoup4==4.12.2
lxml==4.9.3
httpx==0.25.2
aiohttp==3.9.1

# Email & Domain
email-validator==2.1.0
tldextract==5.1.0
dnspython==2.4.2

# Caching
redis==5.0.1
hiredis==2.2.3

# Utilities
python-dotenv==1.0.0
structlog==23.2.0
tenacity==8.2.3

# Monitoring & Performance
prometheus-client==0.19.0

